Remove file after export
