+++ decidim-system/app/controllers/decidim/system/administration_controller.rb
@@ -5,30 +5,12 @@ class Decidim::System::AdministrationController < Decidim::System::ApplicationCo
   end
 
   def backup
-    send_file(dump_database)
+    send_file(export.path) && export.clean
   end
 
   private
 
-  def dump_database
-    time = Time.now.strftime("%F_%H-%M-%S")
-    system(cmd(time))
-    File.absolute_path(file_path(time))
-  end
-
-  def cmd(time)
-    "pg_dump -Fc -h #{host} -d #{database} -f #{file_path(time)}"
-  end
-
-  def database
-    ActiveRecord::Base.connection_config[:database]
-  end
-
-  def host
-    ActiveRecord::Base.connection_config[:host]
-  end
-
-  def file_path(time)
-    Rails.root.join("backup_#{database}_#{time}.psql")
+  def export
+    Decidim::System::DatabaseExport.new
   end
 end
