+++ decidim-system/app/jobs/decidim/system/clean_database_export_job.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Decidim
+  module System
+    class CleanDatabaseExportJob < ApplicationJob
+      queue_as :default
+
+      def perform(export_path)
+        File.delete(export_path) if File.exist?(export_path)
+      end
+    end
+  end
+end
