+++ decidim-system/lib/decidim/system.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require "decidim/system/engine"
+require "decidim/system/database_export"
 
 module Decidim
   # This module contains all the logic related to a system-wide
